Remove customizable for plugins_pip_packages
The pip packages are normally already configurable with the
<image_name>_pip_packages variable. It doesn't make much sense to make
the list of packages installed with the plugins mechanism configurable
via the <image_name>_plugins_pip_packages variable too.

And even if we wanted to, the parameter to the `customizable()`
function should be 'plugins_pip_packages' and not 'pip_packages'.

This commit removes the customizable bits from the plugins install
macro, at least until we get a good use case for it.

Change-Id: I90dd28b8c5d981e6028af353e34645712ef09b0b