Support configuration of Kafka package
This change makes it easier to configure the Kafka package. This
is useful, for example, when installing Monasca which currently
requires an older version of Kafka than that installed here.

The change also fixes an issue with circular symlinks, updates
kafka and verifies the sha512sum of the package. The sha512sum
was taken from here:

https://www.apache.org/dist/kafka/1.0.1/kafka_2.12-1.0.1.tgz.sha512

Change-Id: I79a5a32ee8e5edd915fa88fbc00d0d4ac08b8ebd
Partially-Implements: blueprint monasca-roles