Do not require the bootstrap password to come in via arguments
Currently when this is being ran, and say ara is being used to
capture the running of kolla-ansible ara will capture the full
command line ran (even if no_log is set); because by default these
modules do not hide what they are running.

So to avoid the situation where the command line shows the password
have this also be able to take in the password via an environment
variable as well (which ara will not capture).

Change-Id: I4d42d592d8031d0f3923bccc6b2db1149af08e75