Add network_mode for image building
Add the ability to specify a different network_mode when running the
image builds. Under the covers the default is to use bridge networking
but if that is unavailable or a user would like to specify a different
networking  method they can now do so using the [DEFAULT]\network_mode
configuration.

Change-Id: I1d2398273b57ca4d52bbfacbc896ffecc58bc54b