Clean python requirements from base image
openstack-base image includes a list of dependencies to be
installed in the image. However some of the packages in this
list seems not to be used by any or very few OpenStack services
according to the info found in codesearch.

This patch removes some python packages from this list as we
are in the process of cleaning dependencies in RDO.

Change-Id: I209b16cfd7dff6ce20bc2417a6079fb8597f6774
Closes-Bug: #1766493