Create symlinks for failed builds logs
This allows to know which image failed to build without having to parse
the huge job-output.txt file.

This commit creates a convenient 000_FAILED_image_name.log symlink to
the relevant log files when kolla is configured to output to a log
directory. This creates symlinks only for the parent image with error,
i.e. it skips images that failed to build because of parent error.

Change-Id: I05954e2335dc3fac469276809c91f9c87e3ecab2
(cherry picked from commit 58742c212e0bdc5092ba6cc0da2fc8c5589bb172)