Stop showing passwords when bootstrap script is ran
Currently this causes bash to echo all lines parsed
and executed; which makes it not so nicely output the
bootstrapping password.

This is not something we should encourage and have show
up in peoples logs or other so stop doing that.

Change-Id: Iac963a5df393d0359b4c8f93b8756ca168f6f193
Closes-Bug: #1775468
(cherry picked from commit 3241012158b12ae578473ee6aa6c414d7443374c)