Remove useless retrieval of karaf for ODL image
When building images with type=source, the opendaylight would retrieve
a 400MB+ archive for distribution-karaf-0.6.2-Carbon.tar.gz and did
absolutely nothing with it since the ODL image installs the necessary
bits from distro packages unconditionally.

This commit removes the retrieval of the karaf archive. This should
help make the gate more stable as we've seen the build timing out when
retrieving the karaf archive.

Change-Id: I0e31d91b4e1ad4ae4ec5c25847c7798724ed9eea
Closes-Bug: #1774851