Show disk space used by docker images in CI
The `docker system df` command can show the total size of the docker
cache. If you pass it the `-v` option, it will break it down per image,
replacing the use of `docker ps -a` and `docker images`.

This should be very useful in CI to see the variation in the images
size.

Change-Id: I407434d88f59df65693974265551c624420d4aaf