Remove Monasca statsd client Docker files
The monasca-statsd repo is a statsd client library with Monasca
extensions such as support for adding dimensions to metrics. It is
designed to be included in application code for sending application
specific metrics to the Monasca statsd service, which is part of the
Monasca Agent.

This commit removes the Monasca Statsd Docker image because the normal
use case would be to include it as a library in an application. The
monasca-statsd tarball is left so that it can be added to containers
using the additions mechanism if a user wants to generate application
specific metrics.

Partially-Implements: blueprint monasca-containers
Change-Id: I20ec284e8f51eb046cc6de2636680847097a3ac1