Stop manually copying sahara-dashboard config files
RDO review https://review.rdoproject.org/r/#/c/14081/ changed the
locaction of the config files for the openstack-sahara-ui and added the
symlinks to be more consistent with the other horizon plugins packages.
It's not necessary anymore to treat sahara-ui as a snowflake and the old
copy in horizon dockerfile must be removed.

Change-Id: Ib222809ee67d103469e9d467727e0e45eaac77e8
Closes-Bug: #1784233