Using qemu-kvm on rhel
qemu-kvm-ev was used with centos,
in case of rhel when you have
access to the virtualisation related channels qemu-kvm pulls
qemu-kvm-rhev which is the recommended one.

qemu-kvm-ev was referenced by devstack long time ago,
but the explicit package install on rhel family today
is qemu-kvm and not qemu-kvm-ev.

I am unsure about the other distros, so just changing rhel.

Change-Id: If6d1e0367730fa80284395be33674a7cd1465389