Add monasca user to Monasca thresh container
The monasca-thresh container is special in that inherits from
the Storm container so that it has access to the Storm client.
We could make monasca-base inherit from Storm, but it would
bloat the other Monasca images which may not run on the same
nodes as the Storm topology.

Partially-Implements: blueprint monasca-containers
Change-Id: Idd1703b72e45bb9d124e06a28738b18da0bc6918