[CI] Support building source images with in-review changes
Following this change it is possible to add 'Depends-On' lines to a
kolla commit message that point to reviews in other projects. For
example, this could be used to test building and deploying container
images containing a change to the nova project that is currently in
review.

It also allows to introduce jobs testing against Kolla in other
projects.

The kolla-build-config Ansible role can also be used by other projects,
such as kolla-ansible, to generate the necessary configuration.

This takes us closer to the Kolla Builder - Next Generation concept [1].

[1] http://lists.openstack.org/pipermail/openstack-discuss/2020-April/014255.html

Change-Id: Ia2efa6b5d7278b75ab8dec23207f92d6bd7c58f0
Co-Authored-By: Radosław Piliszek <radoslaw.piliszek@gmail.com>
(cherry picked from commit 6bfe9b8749c9a41b3d37816879bb0fac1fdd60d5)