Fixed file modes in toolbox:/usr/share/ansible
Sometimes, when deploying with user root, the local repo has files
of mode 0600. This has been found to cause "Permission denied"
errors when deploying Glance.

This little fix is to explicitly set a more permissive file mode on
those files to allow the deployment of Glance to setup it's service
and endpoints.

Change-Id: I30ea6898e7bf1b6713cb41656420e8d515be224e
Signed-off-by: Gary Perkins <gary@garyperkins.com>