Add Monasca Grafana image
This provides support for building the Monasca fork of Grafana
which features Keystone integration.

It is primarily for use with the Monasca Grafana datasource, although
other datasources can be installed if required.

In the future it would be more efficient to build and host the packages
externally.

Implements: blueprint monasca-grafana
Change-Id: I6cb175868d1993ae31700de36b721f4833c4164b