Remove tgtd and scsi-target-utils support in CentOS/RHEL 8
In CentOS/RHEL 8 there is no scsi-target-utils package, nor is it
available in EPEL. In RHEL 7 and beyond the LIO kernel subsystem can be
used instead of the tgtd daemon.

This change removes support for the SCSI target daemon on CentOS/RHEL 8.
The 'tgtd' image is no longer available for CentOS/RHEL 8.

Change-Id: I56f230d66f75dd0546325676278f91579f08c822
Signed-off-by: Maciej Kucia <maciej@kucia.net>
Co-Authored-By: Maciej Kucia <maciej@kucia.net>
Co-Authored-By: Mark Goddard <mark@stackhpc.com>
Related: blueprint centos-rhel-8
Depends-On: https://review.opendev.org/631479/