Add tripleo container build job
The tripleo continainer build process consumes the kolla project. We
would like to add a non-voting job for kolla as the start of better
integration between the two.  This job runs the tripleo build process
which calls kolla-build under the covers.  Failures in this job might
indicate breakages in the kolla build api.

Change-Id: I5ce7c5888858f535eee0b5539c1e2d7978a7438e