Use the base OS image for tripleoclient base
Building from openstack-base brings a lot of unneeded overhead into the
end container. We only need a minimal setup, so let's rely on
dependencies of the python-tripleo package instead of explicitly
pulling in world from openstack-base. That saves space, and reduces
potential attacks vectors and CVEs to maintain for the tripleoclient
container.

Related: blueprint zero-footprint

Change-Id: If70351aac67bf62274a068cb4ebfe9c5270da5ae
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>