Add distro_python3 option
As we add support for python3 on RHEL based systems, we need to add a
flag to key off of in the Dockerfiles to swap out the python3 package
names.  This change adds a distro_python3 configuration option that can
be defined to force python3 or python2 package names.  If this option
is not provided, we set this to True for RHEL 8 based distros and false
for other distros.

Change-Id: I4028991bad92c0e8e21066cc4173c06ce5eba393
Related-Blueprint: python3-support