Add base_package_type
Currently in all the Dockerfile.j2 files, we are doing base distro if
conditions for package names when installing via a binary.  Rather than
duplicating these base_distro in [<distros>], let's provide a single
value for determining the binary installation method. This change adds a
base_package_type which we can use to lookup if 'rpm' or if 'deb' rather
than relying on specific distro names.

Change-Id: I8a0c6c8fbc32a8cfa6932fddf28a449fceda3d49