Sahara: install also the split plugins
The code of Sahara plugins has been extracted during the Stein cycle.
Add support for:
- the source-based deployment method (a beta release of
  the plugins is available on pypi) and
- the RPM-based method (RDO contains the packages for the plugins.)

This change is required also for the RPM method. Even if
the RPM packages of the plugins are linked through weak dependencies
to the sahara-api and sahara-engine packages (Supplements),
weak dependencies are currently disabled here.

Change-Id: I5241d84aa1ca2830cfdb60c1a3ff4e958a21b345