Write hash after compressing horizon static assets
If you run the horizon container with a restart policy of never, and the
static asset compression fails, the container fails to start, e.g. (seen
in CI)

++ /usr/bin/manage.py collectstatic --noinput --clear
Traceback (most recent call last):
  File "/usr/bin/manage.py", line 18, in <module>
      from django.core.management import execute_from_command_line
  ImportError: No module named django.core.management

However, with a restart policy of unless-stopped (the default).  the
horizon container will fail the first time but will succeed after a
restart. This is because the settings bundle hash is updated before the
compression is attempted.

To fix this, update the hash after compression completes successfully.

Change-Id: I3f1682ed55e7d45c1e97a1b34c7993aa839f0bb2
Closes-Bug: #1824137