Fix multiple mon deployment failures
When 4 or 5 mons are deployed, all mons will always be in the election
state, and the ceph cluster cannot be accessed.

After comparing with ceph-deploy, I found that when the cluster was
initialized, it did not use monmap. When adding a new mon node, it
needs to use monmap. According to the above situation, I fixed the
problem.

Change-Id: Idbfc6741d371adcb06ed97d578d75d3bce862743
Closes-bug: #1824783