Fix the problem of osd initialization failed
When deploying osd, if the user does not use the extra block
partition, the kolla will automatically partition the disk and then
clean up the data on the disk partition. Sometimes the disk partition
will not be updated, there will be an error not finding the partition.

This commit fixes the problem.

Change-Id: I14708f38614dcb75268c2f460ae3d921748c2d10
Closes-bug: #1824787
(cherry picked from commit ddab09fdd8d7c08ce8db70948ea12571bb4267a8)