Add python-sqlalchemy-collectd client
sqlalchemy-collectd was added in
I2bcaa44c184bb24e5caeeb8d204dda627de59c91, however it also must
be installed alongside anywhere that python-sqlalchemy is installed
since it provides an in-process plugin that transmits metrics
to a remote collectd daemon.

Also adds the python3 version of the package to the py3 version of
the collecd/Dockerfile.j2.

Note this is currently an *RPM only* package, and as of this
commit is not yet availble for other packaging systems.

Change-Id: If598da717653a383a2d3b3373c56517f8bca832f