Use td-agent-gem to install fluentd monasca plugin on ubuntu
If monasca is enabled on ubuntu, we see the following error in fluentd:

[error]: config error file="/etc/td-agent/td-agent.conf"
error_class=Fluent::ConfigError error="Unknown output plugin 'monasca'.
Run 'gem search -rd fluent-plugin' to find plugins"

This is because we use 'fluent-gem install' to install the gem, which
works on CentOS but on Ubuntu it is necessary to use 'td-agent-gem
install'.

Change-Id: I515f9764ca93a41a871e520696193a9298856d08
Closes-Bug: #1830147