Remove HTTP configuration shipped in openstack-nova-placement-api RPM
The RDO openstack-nova-placement-api RPM ships with an HTTP
configuration deploying placement on port 8778, while Kolla-Ansible
creates another file to configure it to listen on port 8780. If
`placement_api_port` is changed to use port 8778, the placement_api
container will fail to start because the port is already in use.

This commit removes the configuration file shipped with the RPM package
to avoid this conflict.

Change-Id: I7babdb794bec3b08047101e2537f030e2217cedb
Closes-Bug: #1830420