Fix ownership of ironic inspector DHCP hostsdir
If you run with enable_ironic set to true and
ironic_inspector_pxe_filter set to dnsmasq (the default in stein),
ironic inspector can fail with the following in the logs:

Permission denied: u'/var/lib/ironic-inspector/dhcp-hostsdir/<MAC address>'

A node must be registered with a port for this to happen.

Weirdly this happens on centos/source, ubuntu/source, and ubuntu/binary,
but not centos/binary.

This change changes the ownership of
/var/lib/ironic-inspector/dhcp-hostsdir to ironic-inspector user to make
it writeable.

Change-Id: I19447727f19dbd9c0a3e17d218b48ddc4c253587
Closes-Bug: #1832026
(cherry picked from commit 41bcd7b4961fd52066b79804fdce9f81cc129735)