Fix lower-constraints test fail due to failing dep build
Ubuntu Bionic updated OpenSSL package to new release:
from 1.1.0 to 1.1.1
It caused cryptography in version from lower-constraints
to fail to build in Zuul.
Additionally, constrained cryptography version did not
match the requirement for constrained PyOpenSSL.
This commit bumps the version to the minimum required
by constrained PyOpenSSL.
It works because manylinux1 (binary) wheel exists.
This version agrees also with upper-constraints for Queens.

Dependency on cryptography is removed as it is not used directly
by kolla (and requirements checker requires lower-constraints
matching version - let's not create an illusion of usage).

Change-Id: I8ae19f212049cae5dfb4e8dd6ec6cd14a8c46e0e
Signed-off-by: Radosław Piliszek <radoslaw.piliszek@gmail.com>