Bump Monasca Fluentd output plugin
This resolves an issue where the previous version of the plugin
would fail to re-send buffered logs if a post to the Monasca Log
API failed. With this new version of the plugin, logs are correctly
buffered and resent.

Change-Id: Ic1d93a8ea5f2571edfacf50809d51a8b77959a73
(cherry picked from commit d7e6c28272a2efddec8a1747067e4fdee4ff50ba)