Add more entries into SKIPPED_IMAGES
Did build of all images for AArch64, looked at failures and here is a
list of ones which never build due to lack of packages.

There are some images which could get to the list but there are already
patches for them in Gerrit.

Also added three more entries for Debian due to missing packages.

Change-Id: Ie259a38aad84bd8f5392f19791edfbd1f8ec4b1a