Move pike jobs to use validated RDO Trunk
Pike is EM now which means that projects are not creating new releases.
For RDO this means official repos are not considered supported anymore
and may be removed at some point.

The recommendation for CI is to use RO Trunk repo for pike which follows
the stable/pike branches instead of last releases.

Change-Id: If5670473b69578f2a35646a54be9be88aee6bbb2