Add partition detection to fix osd initialization failure
The changes in the following commit did not completely solve the
problem of osd initialization failure:
https://review.opendev.org/652612

In order to solve this problem, the partprobe command consistent with
ceph-disk is added, and then a loop is added to detect the partition,
which is acquired every 1s. When the partition appears, the osd
initialization is continued.

Change-Id: I0ca255c6358132d9e3acfa6b610b70a78756512c
Closes-bug: #1824787
(cherry picked from commit 342c9f0cd0dc4b97870dc40fab6c5a7d241b40ed)