Add prometheus_client to monasca-agent image
The prometheus_client package is required to use the Prometheus plugin.
It makes sense to include this dependency by default since Kolla Ansible
can deploy many Prometheus exporters, making it easy to scrape their
metrics and store them in Monasca. The increase in size of the container
image is negligible.

Change-Id: I9b0a162513ed436930c4541b758fc9cb2ef97e96