Use openstack-tempest-all for rpm tempest container
openstack-tempest shipped from RDO does not have the tempest
plugins, It has just tempest cli and library.

https://review.opendev.org/#/c/651601/ updated the rpm package
name while solving different issue and TripleO stein CI,
most of the tempest tests are getting skipped as tempest plugins
are not there.

openstack-tempest-all is the meta package contains all the plugins,
and installing that fixes the issue.

Change-Id: Ifb616b082c9d82f4736e4b688cafc3a58a0c7a9f
Signed-off-by: Chandan kumar <chkumar@redhat.com>