Fix retries when pushing images
Currently the retry mechanism is broken for pushing because the image
state gets set to an error state, and is never reset to 'built'. This
prevents the PushTask from setting success to True.

This change sets the image state to 'built' if the push succeeds,
ensuring it overrides any previous failures.

Change-Id: I93fc0e383da8fec6b3ca31f8094321c2a0c3af71
Closes-Bug: #1844697