ovn: Create the ovn dbs only if the schema file exists
OVN has been split out from openvswitch and it's code is here [1].
With the split we have changed the ovn runtime/data dirs to
/etc/ovn and /usr/share/ovn respectively.

If a user uses latest OVN (when used after the split from
openvswitch), then the extend_start.sh of ovn-nb-db-server and
ovn-sb-db-server will fail because the OVN schema files will be in different
location (new location is /usr/share/ovn/ovn-nb.ovschema and /usr/share/ovn/ovn-sb.ovschema)

The old location is - /usr/share/openvswitch/ovn-nb.ovschema and
/usr/share/openvswitch/ovn-sb.ovschema

This patch checks for the presence of the schema files in the old location before
accessing them.

For latest OVN, it is advisable to use the "ovn-ctl" script to start
and stop OVN ovsdb-servers. ovn-citl script takes care of creating the dbs
and updating the dbs if the schema is updated.

[1] - https://github.com/ovn-org/ovn
      https://github.com/openvswitch/ovs/commit/f3e24610ea18eb873dc860f1710432e9aacd27fd

Change-Id: If2bab2fac5be28ae7e1cd02de13a2de94fe17bac
Closes-bug: 1849633
Signed-off-by: Numan Siddique <nusiddiq@redhat.com>