Use ovn-ctl script to start OVN ovsdb-servers
Using the ovn-ctl script to start the OVN ovsdb-server is the better
way. This script is used to start and stop all OVN systemd services.
This script takes care of creating the database from the schema,
updating the database from the schema when the schema version gets
updated.

The present scripts - start_nb_db_server.sh and start_sb_db_server.sh
doesn't take care of schema updates.

With this, we don't need extend_start.sh in ovn-nb-db-server/ovn-sb-db-server
to create the database from the schema.

This patch also deletes kolla_ovn_extend_start from the ovn-base image as it
is not required.

Closes-bug: 1850025
Change-Id: I31a372d801dace92950cd85b29b308431f78bbdb