Switch to stable/train for source images on CentOS/RHEL
OpenStack projects are beginning to drop python 2 support on the master
branch. CentOS 8 is not yet ready, so our best option to continue
building and testing CentOS is to use the Train tarballs for source
images.

This change also introduces a new openstack_branch config option which
can be used to set the branch used for source images. For CentOS/RHEL,
the default is stable-train, for Debian/Ubuntu it it master.

Change-Id: Ieab5e06a3c05951d97bb979ab735b4cef675cc0a