CentOS 8: Remove shellinabox from ironic-conductor
shellinabox is used by ironic-conductor to provide a browser-based shell
to access the consoles of nodes. It is not used by all console drivers,
and is not used in the integration with nova serial consoles.

The package was previously installed from EPEL7, but is not currently
provided by EPEL8.

This change removes the package to allow the ironic-conductor image to
build.

Change-Id: I210d9fddd6948dab1feb5abf59a50bce86f21f0f
Partially-Implements: blueprint centos-rhel-8
(cherry picked from commit 397651ec4b5b93a55d0b8b8f69dfb37dd85cfc8d)