CentOS 8: disable repo_gpgcheck for grafana image
DNF does not seem to accept keys imported by rpm --import for signing a
repo (as opposed to packages). This causes it to prompt during package
install, which does not work without a terminal in the build container.

We have gpgcheck=1 for the grafana repo, so packages are at least
signed. We don't have repo_gpgcheck elsewhere, so it seems reasonable to
disable it for grafana.

Change-Id: Ice120e8001fd9e8bb37ca4f2df26259cf798ee0e
Partially-Implements: blueprint centos-rhel-8