Throw TypeError in repository enable/disable macros
Currently we do not validate the type of the reponames argument to the
enable_extra_repos and disable_extra_repos macros. It should be a list,
but we merged a call where the argument was a string, which interpreted
as a list of single characters ('e', 'p', 'e', 'l'), and silently
ignored.

This change validates that the reponames argument is a list to avoid
this issue.

Change-Id: I497c8b2c1dc548e678698474d3604664cf8eeb07