Fix invalid test in cinder data migration script
The extend_start.sh script for cinder-api uses an incorrect test to
check if the MAX_NUMBER variable is set. This leads to a incorrect
cinder-manage command being run if this variable is not set, making the
database migration fail.

This change fixes the test to properly use the MAX_NUMBER variable.

Change-Id: Ie46da709db0de09a9cf641b3c154275282f213fc
Closes-Bug: #1866827