Sensu: Remove ceph and process-checks plugins from client
The sensu image became unbuildable, failing to install the ceph plugin:

    Unable to resolve dependencies: sensu-plugins-ceph requires english (= 0.6.3)

This has happened because the author of the 'english' gem has seen fit
to publish a new 0.7.0 release, and 'yank' [1] all previous releases
from rubygems.

This change removes the ceph and process-checks plugins, which both
suffer from a dependency on the yanked english 0.6.3.

[1] https://rubygems.org/gems/english/versions/0.6.3

Change-Id: I9d34413fa10b5a2920aca60b08141bfaddcc7415
Closes-Bug: #1871600