switch to official CentOS 8 repositories
RDO builds of stable/train for CentOS 8 went into phase when we do not
need to use delorean builds.

EPEL is disabled like it is in Ussuri (RDO packages may conflict with
it).

Erlang for RabbitMQ is used from CentOS 8 'rabbitmq-38' repo while
RabbitMQ is still 3.7.* from same place as usual.

'python3-pika' is dropped from openstack-base as it is EPEL only package
and is not used by any image (dropped in Ussuri in
https://review.opendev.org/#/c/710003/ patch)

AArch64 is not supported until CentOS 8.2 release.

Change-Id: I9d3ed54f4d10ef7ad2d52ed3aeca15f3968cd163