Always build explicitly matched parents
If --skip-parents is used, then _any_ parent is skipped in the build.
However, there are cases where you want to rebuild not just a leaf node,
but also a parent node or two. For example a filter could be "^nova-",
requesting all Nova images. To properly rebuild the base image
(nova-base), we need to ensure that it is built, even though technically
it is a parent (of e.g., nova-api).

Change-Id: I34c031ed8308fd625c74b55ea3052608c1f653eb