Switch to tarballs.opendev.org for source images
OpenStack tarballs are hosted on tarballs.opendev.org.
tarballs.openstack.org is just a redirect.

This change switches the tarballs-base configuration option to
https://tarballs.opendev.org. Since the new site is namespaced, we need
to add openstack/ to the URL path, or in some cases x/.

The following projects now publish under the x/ namespace:

* networking-ansible
* networking-mlnx
* vmware-nsx
* vmware-nsxlib

This reverts commit 1c2ee4993d824ed4dc5331706a86491a459c3c4e, adapting
it to the master branch instead of stable/ussuri.

Change-Id: If9bd73a2e758c9da87f9a9a60fe075d807e3ca00
Co-Authored-By: Mark Goddard <mark@stackhpc.com>