Add Logstash 6 for Monasca on Centos 8
On Centos 8 we have Elasticsearch and Kibana 6, but Logstash (used by
Monasca) is stuck on 2 which is not compatible. See the product
compatibility matrix [1]. This adds Logstash 6, so that Monasca
can use the full ELK 6 stack on Centos 8.

Note that Monasca doesn't officially support ELK 6, but it does work.

[1] https://www.elastic.co/support/matrix#matrix_compatibility

Partial-Bug: #1884090
Change-Id: Id6f79e301cc109a029913f5628bab1848c138e7f