Fix ceilometer-upgrade params
It seems [1] has introduced a slight regression in that it tries
by default to pass an empty string to ceilometer-upgrade which
makes it fail:

ceilometer-upgrade: error: unrecognized arguments:

(note there is nothing more on that line)

This patch fixes that by allowing the params to be of any number,
including zero.

Backport to Ussuri.

[1] 3d587409ee068843cfa28710f09f65af48b69367

Change-Id: I989a1dd30d1656b5c1fda7a46abbf5f265fbb72c
Closes-Bug: #1884919