Remove weird deps
This is a follow-up on "Refactor httpd install to base image"
[1].

It seems a copy-paste algorithm was used to craft Dockerfiles
for some httpd-enabled services which resulted in an abundance of
ldappool packages getting installed, even in the 'source' case.
This seems to have also kept ldappool at a lower version because
it did not get updated via pip later.
This patch deals with that and also moves ldap deps for Keystone
to their proper place in 'source' case (extras).
Note Keystone client gets installed in openstack-base.

Cinder does not need to include Keystone either.

[1] https://review.opendev.org/744037

Change-Id: I017d7a6a5d2b1ae6c04556dcf172453a36de5be7